// src/app/models/unit.ts
export interface Unit {
    id: number
    itemUnitNo: number
    itemUnit1: string
    itemUnitHindi: string
    rateOfUnitinKgs: number
    percentComission: number
    weightCuttingPerUnit: number
    hammaliPerUnit: number
    tulaiPerUnit: number
    khadiKariPerUnit: number
    weightCuttingPerUnit1: number
    hammaliPerUnit1: number
    tulaiPerUnit1: number
    khadiKariPerUnit1: number
    weightCuttingPerUnit2: number
    hammaliPerUnit2: number
    tulaiPerUnit2: number
    khadiKariPerUnit2: number
    weightCuttingPerUnit4: number
    hammaliPerUnit4: number
    tulaiPerUnit4: number
    khadiKariPerUnit4: number
    fkBranchId: number
    isActive: string
    createBy: number
    createDate: string
    modifyBy: number
    modifyDate: string
    sysdate: string
  }